clean up + dual platform

+++ install.sh
@@ -1,13 +1,30 @@
 #!/usr/bin/env sh
 
+# Get the os parameter, if not provided, ask for it (macos or debian)
+if [ -z "$1" ]; then
+  echo "Please provide the OS type (macos or debian)."
+  echo "Usage: $0 <os_type>"
+  exit 1
+fi
+
+OS_TYPE="$1"
+
+if [ "$OS_TYPE" != "macos" ] && [ "$OS_TYPE" != "debian" ]; then
+  echo "Invalid OS type. Please use 'macos' or 'debian'."
+  exit 1
+fi
+
 # Current path
 DOTFILES="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 # ZSH
 ln -sf $DOTFILES/.zshrc $HOME/.zshrc
 
-# Fira Code Nerd Font
-brew install font-fira-code-nerd-font
+# Only for macOS
+if [ "$OS_TYPE" = "macos" ]; then
+    # Fira Code Nerd Font
+    brew install font-fira-code-nerd-font
+fi
 
 # Neovim
 rm -rf $HOME/.config/nvim
@@ -17,24 +34,14 @@ ln -s $DOTFILES/nvim $HOME/.config/nvim
 rm -rf $HOME/.config/phpactor
 ln -s $DOTFILES/phpactor $HOME/.config/phpactor
 
-brew install watchman
+if [ "$OS_TYPE" = "macos" ]; then
+    brew install watchman
+else
+    apt install -y watchman
+fi
 
 # Lazygit
 ln -s $HOME/.config/lazygit/config.yml ~/Library/Application\ Support/lazygit/config.yml 
-
-# AeroSpace
-rm -rf $HOME/.config/aerospace
-ln -s $DOTFILES/aerospace $HOME/.config/aerospace
-
-# SketchyBar
-rm -rf $HOME/.config/sketchybar
-ln -s $DOTFILES/sketchybar $HOME/.config/sketchybar
-chmod +x $HOME/.config/sketchybar/items/*
-chmod +x $HOME/.config/sketchybar/plugins/*
-
-# Borders
-rm -rf $HOME/.config/borders
-ln -s $DOTFILES/borders $HOME/.config/borders
 
 # Yazi
 rm -rf $HOME/.config/yazi
@@ -46,13 +53,6 @@ ln -s $DOTFILES/karabiner $HOME/.config/karabiner
 
 # Tmux
 ln -sf $DOTFILES/.tmux.conf $HOME/.tmux.conf
-
-# Alacritty
-ln -sf $DOTFILES/.alacritty.yml $HOME/.alacritty.yml
-
-# Kitty
-rm -rf $HOME/.config/kitty
-ln -s $DOTFILES/kitty $HOME/.config/kitty
 
 # Wezterm
 rm -rf $HOME/.config/wezterm
